Backport patch so that 26.1 and 26.2 can use TLS with ELPA
Out of the box, these two versions are configured with an HTTPS URL
for ELPA in `package-archives`, but they cannot download packages from
that location due to
https://debbugs.gnu.org/cgi/bugreport.cgi?bug=34341.

The most sane thing to do in this case seems to be to backport the
patch applied to 26.3, so that they at least work with ELPA out of the
box.

This relates to issue #2.